+++ api/src/main/java/ntk/android/base/entitymodel/article/ArticleContentModel.java
@@ -19,9 +19,9 @@ public class ArticleContentModel extends BaseModuleEntity<Long> {
     @SerializedName("FromDate")
     public String fromDate;
     @SerializedName("Geolocationlatitude")
-    public long Geolocationlatitude;
+    public Double Geolocationlatitude;
     @SerializedName("Geolocationlongitude")
-    public long Geolocationlongitude;
+    public Double Geolocationlongitude;
     @SerializedName("Keyword")
     public String Keyword;
     @SerializedName("LinkFileIds")
